Don't combine modules if root node is already set (#22)
This PR introduce a new default behaviour to the `CombineModulesVisitor`. If, when visiting the `RootKorok`, we see that a `RootNode` has already been set up, we stop trying to combine modules. This provides ways for plugins to provide their own logic for propagating information up the tree without having the final hardcoded step overriding their work.

This PR also introduces a new `force` option for that visitor to override the existing `RootNode` anyway.