Provide no-op macros in Solana targets (#38)
This PR removes all dependencies and logic in the `codama-macros` crate when it is imported in a Solana environment (using the `"solana"` target).

This means, when building a program via `cargo build-sbf`, the `codama-macros` crate won't have any work to do before expanding the macros to nothing and it won't even import any dependencies.